pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datakit"
version = "0.1.0"
description = "Tools to merge and remap YOLO-format computer vision datasets"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
  "rich>=13.7",
]
authors = [
  { name = "Imtiaz Ahammed Anik" }
]

[project.optional-dependencies]
visualize = [
  "matplotlib>=3.9.4",
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.scripts]
datakit = "datakit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
